+++ pMuTT/chemkin/__init__.py
@@ -46,7 +46,7 @@ class CatSite:
                 JSON representation
         Returns
         -------
-            BaseThermo : BaseThermo object
+            EmpiricalBase : EmpiricalBase object
         """
         json_obj = remove_class(json_obj)
         return cls(**json_obj)