+++ pMuTT/io_/jsonio.py
@@ -58,7 +58,7 @@ def type_to_class(class_str):
     from pMuTT.eos import IdealGasEOS, vanDerWaalsEOS
     from pMuTT.reaction import Reaction, Reactions
     from pMuTT.reaction.bep import BEP
-    from pMuTT.empirical import BaseThermo
+    from pMuTT.empirical import EmpiricalBase
     from pMuTT.empirical.nasa import Nasa
     from pMuTT.empirical.shomate import Shomate
     from pMuTT.empirical.references import Reference, References
@@ -78,7 +78,7 @@ def type_to_class(class_str):
         "<class 'pMuTT.reaction.Reaction'>": Reaction,
         "<class 'pMuTT.reaction.Reactions'>": Reactions,
         "<class 'pMuTT.reaction.bep.BEP'>": BEP,
-        "<class 'pMuTT.empirical.BaseThermo'>": BaseThermo,
+        "<class 'pMuTT.empirical.EmpiricalBase'>": EmpiricalBase,
         "<class 'pMuTT.empirical.nasa.Nasa'>": Nasa,
         "<class 'pMuTT.empirical.shomate.Shomate'>": Shomate,
         "<class 'pMuTT.empirical.references.Reference'>": Reference,
@@ -100,17 +100,19 @@ def type_to_class(class_str):
 
 
 def remove_class(json_obj):
-    """Removes the 'class' entry from the JSON object.
+    """Removes unnecessary entries from the JSON object when reinitializing the
+    pMuTT object
 
     Parameters
     ----------
         json_obj : dict
-            JSON object with 'class' entry
+            JSON object unnecessary entries
     Returns
     -------
         json_obj : dict
-            JSON object without 'class' entry
+            JSON object without unnecessary entries
     """
     json_obj.pop('class', None)
+    json_obj.pop('type', None)
     json_obj.pop('_id', None)
     return json_obj